(breaking) Change CompletingCommandExecutor.complete() to take cursor instead of argNumber and offset
This is because argNumber (ambigous name btw) becomes invalid after
parsing flags. Before the change executors who use flags have to convert
the pair of numbers to cursor, parse flags, and then convert it back.
This is not straightforward and may be confusing.

After the change the completer will be given raw cursor position which
never becomes invalid (because it doesn't depend on the current index of
CommandArguments). They will call CommandArguments.offsetToArgument()
when they need to know which argument is being attemted to be completed,
and pass the raw cursor everywhere else.