Add CommandArguments.completeRemainingStrings()
Also:
* move a part of complete() to a new method completeRaw() which supports
completion of not-necessarily-the-current-argument, but needs the caller
to provide the raw start string
* add the argName != null condition to hasOverride and remove it from
currentArgument's override condition

popRemainingStrings() pops more than one argument/word, so it needs
custom completion method, like flags. Regular complete() won't work,
except (accidentally) for the closing quote. Add
completeRemainingStrings() as the completion method for
popRemainingStrings().

Before the change, currentArgument() checks for (argName != null &&
hasOverride(argName)). This makes it inconvinient for other methods to
check for overrides, because they have to do the null check, too. Adding
the check inside hasOverride() and removing it from currentArugment()
makes it a lot easier to check if there's an override - it's enough to
just call hasOverride().